moved parts to their own directories
