--- config/vehicle.ini
@@ -1,29 +1,0 @@
-#Settings for the car. 
-[vehicle]
-id=mycar
-loop_delay=.08  #Forced delay (seconds) between web requests.   
- 
-[camera]
-loop_delay = .08 #Forced delay (seconds) between refreshing the camera.
-
-#Channel and PWM settings for the motor.
-[throttle_actuator]
-channel=0
-max_pulse=500
-zero_pulse=370
-min_pulse=220
-
-#Channel and PWM settings for the steering servo.
-[steering_actuator]
-channel=1
-left_pulse=460
-right_pulse=260
-
-#Where your local pilot lives.
-[pilot]
-model_path=~/mydonkey/models/default.h5
-
-#folders to save data and configs
-[mydonkey]
-path = ~/mydonkey
-folders = sessions,models,results,datasets
