--- docs/01-build_a_car.md
@@ -1,3 +1,0 @@
-# Build a donkey vehicle. 
-
-Anyone can build their own robocar that runs Donkey. Once you have the parts it will take ~4hrs to start driving. See the [detailed instructions](https://docs.google.com/document/d/11IPqZcDcLTd2mtYaR5ONpDxFgL9Y1nMNTDvEarST8Wk/edit).
