--- docs/hardware/README.md
@@ -1,16 +1,0 @@
-##Mechanical Hardware Documentation
-The hardware for the Donkey is fairly straight forward, however having consistent hardware is important.  It is to everybody's benefit that we end up with same/similar training data. Only two cars are supported right now:
-
-*	Helion Conquest from Hobbytown
-* 	Trooper from Hobby King (both Pro and non Pro version are supported)
-
-We will evaluate support for other cars on an ongoing basis.  When purchasing the car having a Remote control is not required so an "ARR" version might be less expensive.
-
-###Top Plate
-The top plate is two parts laser cut from 3mm material like baltic birch plywood.  the two top plates are then glued together.   
-
-###Roll Cage
-The roll cage is an optional part printed on a 3D Printer, it can be attached to the top plate using M3 screws <note: provide length>.  While optional the roll cage does help protect electronics, but possibly more importanly provides much needed handles for the car such that you dont touch wiring connections when moving the car. 
-
-###Camera Holder
-The adjustable camera holder has 3 positions 0, -15 and -30 degrees.  The -15 degree setting is recommended for all data collection, other settings are available for testing.  Both camera holders are held down with M3 screws and optionally can use a square nut to for easier tightening.
