+++ docs/stores.md
