+++ donkey/parts/__init__.py
@@ -1,0 +1,16 @@
+
+
+from .controllers.web import LocalWebController
+
+from .sensors.cameras import PiCamera
+from .sensors.lidar import RPLidar
+
+
+from .pilots.keras import KerasCategorical
+
+from .stores.tub import Tub
+
+from .transforms import Lambda
+
+from .simulations import SquareBoxCamera
+from .simulations import MovingSquareTelemetry
