+++ donkey/parts/controllers/templates/session_list.html
