+++ donkey/parts/controllers/templates/static/main.js
