+++ donkey/parts/controllers/templates/static/nipple.js
