+++ donkey/parts/controllers/templates/vehicle.html
