+++ examples/web_controller.py
@@ -16,13 +16,13 @@ V.mem.put(['square/x', 'square/y'], (100,100))
 
 
 #display square box given by cooridantes.
-cam = dk.sensors.SquareBoxCamera(resolution=(200,200))
+cam = dk.parts.SquareBoxCamera(resolution=(200,200))
 V.add(cam, 
       inputs=['square/x', 'square/y'],
       outputs=['square/image_array'])
 
 #display the image and read user values from a local web controller
-ctr = dk.controllers.LocalWebController()
+ctr = dk.parts.LocalWebController()
 V.add(ctr, 
       inputs=['square/image_array'],
       outputs=['user/angle', 'user/throttle', 'user/mode'],
@@ -30,7 +30,7 @@ V.add(ctr,
 
 #transform angle and throttle values to coordinate values
 f = lambda x : int(x * 100 + 100)
-l = dk.transforms.Lambda(f)
+l = dk.parts.Lambda(f)
 V.add(l, inputs=['user/angle'], outputs=['square/x'])
 V.add(l, inputs=['user/throttle'], outputs=['square/y'])
